Make mod4 layer work fully with Noted
Noted and Mine use another key for the mod3 layer than the "classic" Neo(2), Bone, and NeoQwertz.
Also the keys arount those layer keys on the very right side of the keyboard work slightly differently from each other on the level 4 (the level with the navigation keys and virtual num block) I'm talking about the keys on the right of this virtual numblock.
I really wanted to make those keys work in a consistent manner. But this would mean that this change in these complex karabiner modification rules, would break the currently used keylayout files.
To make it consistent, both these rules and the keyfiles must be updated simultanously.
So I decided to make it backwards compatible, meaning my new rules should not break existing Neo, Bone, and NeoQwertz keylayouts.

But now they also work with the keylayout file that I also created on the neo-layout repo. And now it should also be easier to also add a Mine Layout.

This was really tricky! I'm proud of myself for not giving up and finally have found the solution.